+++ k3/k2rps.h
@@ -7,4 +7,4 @@
 #include "custstr.h"
 #include "systimer.h"
 #include "int64voodoo.h"
-void initPlayer(uint64_t N, uint64_t player_type, int priority);
+void initPlayer(uint64_t N, uint64_t player_type, uint64_t priority);
